break out editors into separate files
