+++ public/app/features/transformers/editors/CalculateFieldTransformerEditor/IndexOptionsEditor.tsx
@@ -1,0 +1,30 @@
+import React from 'react';
+
+import { CalculateFieldTransformerOptions } from '@grafana/data/src/transformations/transformers/calculateField';
+import { InlineField, InlineSwitch } from '@grafana/ui';
+
+import { LABEL_WIDTH } from './constants';
+
+export const IndexOptionsEditor = (props: {
+  options: CalculateFieldTransformerOptions;
+  onChange: (options: CalculateFieldTransformerOptions) => void;
+}) => {
+  const { options, onChange } = props;
+  const { index } = options;
+
+  const onToggleRowIndexAsPercentile = (e: React.FormEvent<HTMLInputElement>) => {
+    onChange({
+      ...options,
+      index: {
+        asPercentile: e.currentTarget.checked,
+      },
+    });
+  };
+  return (
+    <>
+      <InlineField labelWidth={LABEL_WIDTH} label="As percentile" tooltip="Transform the row index as a percentile.">
+        <InlineSwitch value={!!index?.asPercentile} onChange={onToggleRowIndexAsPercentile} />
+      </InlineField>
+    </>
+  );
+};
