+++ public/app/features/transformers/editors/CalculateFieldTransformerEditor/constants.ts
@@ -1,0 +1,1 @@
+export const LABEL_WIDTH = 16;
