+++ public/app/features/transformers/editors/CalculateFieldTransformerEditor/index.ts
@@ -1,0 +1,3 @@
+import { CalculateFieldTransformerEditor } from './CalculateFieldTransformerEditor';
+
+export { CalculateFieldTransformerEditor };
